电商认证

+++ service/src/main/java/com/huotu/loanmarket/service/entity/carrier/AsyncTask.java
@@ -38,7 +38,7 @@ public class AsyncTask {
     private String taskId;
 
     @Column(name = "merchant_id")
-    private Long merchantId;
+    private Integer merchantId;
 
     /**
      * 是否失败
